+++ module/data/actor/index.js
@@ -1,0 +1,7 @@
+import {PersonnageData} from './personnage.js';
+import {PnjData} from "./pnj.js";
+
+export const config = {
+    personnage: PersonnageData,
+    pnj: PnjData,
+};