+++ module/data/item/attribut.js
@@ -1,0 +1,13 @@
+const { HTMLField, StringField  } = foundry.data.fields;
+
+export class AttributeItemData extends foundry.abstract.TypeDataModel {
+
+    static defineSchema() {
+        return {
+            description: new HTMLField({ initial: '' }),
+            regles_opt: new HTMLField({ initial: '' }),
+            subtitle: new StringField({ initial: '' }),
+        };
+    }
+
+}