+++ module/data/item/index.js
@@ -1,0 +1,5 @@
+import { AttributeItemData } from './attribut.js';
+
+export const config = {
+    attribut: AttributeItemData,
+};